Move fake hilt module classes
