--- app/src/androidTest/java/com/cryart/sabbathschool/di/FakeRepositoryModule.kt
@@ -1,48 +1,0 @@
-/*
- * Copyright (c) 2021. Adventech <info@adventech.io>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-package com.cryart.sabbathschool.di
-
-import app.ss.lessons.data.di.RepositoryModule
-import app.ss.lessons.data.repository.quarterly.QuarterliesRepository
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.components.SingletonComponent
-import dagger.hilt.testing.TestInstallIn
-
-/**
- * QuarterliesRepository binding to use in tests.
- *
- * Hilt will inject a [FakeQuarterliesRepository] instead of a [QuarterliesRepository].
- *
- * TODO: Move this to a separate test module for re-use
- */
-@Module
-@TestInstallIn(
-    components = [SingletonComponent::class],
-    replaces = [RepositoryModule::class]
-)
-interface FakeRepositoryModule {
-
-    @Binds
-    fun bind(impl: FakeQuarterliesRepository): QuarterliesRepository
-}
